--- mybatis-plus-generator-ui-starter-common/src/main/java/com/github/davidfantasy/mybatisplus/generatorui/common/core/utils/OSUtil.java
@@ -1,24 +1,0 @@
-package com.github.davidfantasy.mybatisplus.generatorui.common.core.utils;
-
-public class OSUtil {
-
-    private static final String OS = System.getProperty("os.name").toLowerCase();
-
-    public static boolean isLinux() {
-        return OS.contains("linux");
-    }
-
-    public static boolean isMacOS() {
-        return OS.contains("mac") && OS.contains("os") && OS.contains("x");
-    }
-
-    public static boolean isMacOSX() {
-        return OS.contains("mac") && OS.contains("os") && OS.contains("x");
-    }
-
-    public static boolean isWindows() {
-        return OS.contains("windows");
-    }
-
-
-}
