--- mybatis-plus-generator-ui-starter-common/src/test/resources/application-test.properties
@@ -1,4 +1,0 @@
-spring.main.banner-mode=off
-logging.level.root=off
-logging.pattern.console=%m%n
-spring.main.lazy-initialization=true
