+++ mybatis-plus-generator-ui-starter-webmvc-api/pom.xml
@@ -7,6 +7,9 @@
         <version>2.0.5</version>
     </parent>
     <artifactId>mybatis-plus-generator-ui-starter-webmvc-api</artifactId>
+    <description>
+        这一层留给用户按照API自己开发接口
+    </description>
     <dependencies>
         <!-- springdoc-common -->
         <dependency>
