--- mybatis-plus-generator-ui-tests/mybatis-plus-generator-ui-actuator-webmvc-tests/src/test/java/com/github/davidfantasy/mybatisplus/generatorui/app2/SpringDocApp2Test.java
@@ -1,29 +1,0 @@
-/*
- *
- *  * Copyright 2019-2020 the original author or authors.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      https://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
-
-package com.github.davidfantasy.mybatisplus.generatorui.app2;
-
-import com.github.davidfantasy.mybatisplus.generatorui.AbstractSpringDocTest;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * 按照Spring Doc标准进行的单元测试
- */
-public class SpringDocApp2Test extends AbstractSpringDocTest {
-
-}
