+++ mybatis-plus-generator-ui-tests/mybatis-plus-generator-ui-actuator-webmvc-tests/src/test/java/test/com/github/davidfantasy/mybatisplus/generatorui/AbstractCommonTest.java
@@ -1,4 +1,4 @@
-package com.github.davidfantasy.mybatisplus.generatorui;
+package test.com.github.davidfantasy.mybatisplus.generatorui;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
