+++ mybatis-plus-generator-ui-tests/mybatis-plus-generator-ui-actuator-webmvc-tests/src/test/java/test/com/github/davidfantasy/mybatisplus/generatorui/TestSqlParser.java
@@ -1,4 +1,4 @@
-package com.github.davidfantasy.mybatisplus.generatorui;
+package test.com.github.davidfantasy.mybatisplus.generatorui;
 
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.collection.CollectionUtil;
