+++ mybatis-plus-generator-ui-tests/mybatis-plus-generator-ui-actuator-webmvc-tests/src/test/java/test/com/github/davidfantasy/mybatisplus/generatorui/app2/SpringDocApp2Test.java
@@ -1,0 +1,10 @@
+package test.com.github.davidfantasy.mybatisplus.generatorui.app2;
+
+import test.com.github.davidfantasy.mybatisplus.generatorui.AbstractSpringDocTest;
+
+/**
+ * 按照Spring Doc标准进行的单元测试
+ */
+public class SpringDocApp2Test extends AbstractSpringDocTest {
+
+}
