--- mybatis-plus-generator-ui-tests/mybatis-plus-generator-ui-actuator-webmvc-tests/src/test/resources/application-test.properties
@@ -1,5 +1,0 @@
-spring.main.banner-mode=off
-logging.level.root=debug
-debug=true
-logging.pattern.console=%m%n
-spring.main.lazy-initialization=true
